+++ src/view/position-app/dialog/draggable/OptionsDraggable.svelte
@@ -1,4 +1,0 @@
-<!-- Provides an empty component for any future `draggable` options -->
-<script>
-</script>
-